Work on tidying up some types and removing dead code
