+++ library/data-model/src/data_storage/queries.ts
@@ -26,50 +26,8 @@ import {
   ProjectUIModel,
   RecordID,
   RecordMetadata,
-  RecordReference,
 } from '../types';
 import {listRecordMetadata} from './internals';
-
-// Get all records of a specific type
-// only referenced in a catch clause in storageFunctions.ts so probably never
-// used
-export async function getAllRecordsOfType(
-  project_id: ProjectID,
-  type: FAIMSTypeName
-): Promise<RecordReference[]> {
-  const dataDB = await getDataDB(project_id);
-
-  const BATCH_SIZE = 100;
-  const records: RecordReference[] = [];
-  let skip = 0;
-  let res;
-  let count = 0;
-
-  do {
-    res = await dataDB.find({
-      selector: {
-        record_format_version: 1,
-        type: type,
-      },
-      limit: BATCH_SIZE,
-      skip: skip,
-    });
-
-    count = res.docs.length;
-    // const hrid = (await getHRID(project_id, o.revision)) ;
-    res.docs.forEach((o: any) => {
-      records.push({
-        project_id: project_id,
-        record_id: o._id,
-        record_label: o._id, // TODO: decide how we're getting HRIDs from db
-      });
-    });
-
-    skip += BATCH_SIZE;
-  } while (res && count === BATCH_SIZE);
-
-  return records;
-}
 
 /**
  * Get an array of records with values that match a regular expression
